requirements.txt
pyqrcode
pypng
pyqrcode
cairosvg
qrtools
pyzbar